Reimplement role-wise discriminativity queries
Now row-wise discriminativity queries are considering a triple
non-discriminative if there's at least one property of pivot
and comparison with the value of feature and a common role.

Prior to the commit a feature would be considered discriminative
if pivot would have at least a single property with the value of
feature and a different role from comparison's feature properties.